requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
ffmpeg-python
PyYAML
